fix(data-quality): use configured datasource tree
